+++ src/script/engine.rs
@@ -42,6 +42,7 @@ pub fn build_engine(defaults: &ScriptDefaults) -> rhai::Engine {
     super::bindings::dict::register(&mut engine, defaults.clone());
     super::bindings::dns::register(&mut engine);
     super::bindings::file::register(&mut engine);
+    super::bindings::hash::register(&mut engine);
     super::bindings::http::register(&mut engine, defaults.clone());
     super::bindings::ldap::register(&mut engine, defaults.clone());
     super::bindings::memcached::register(&mut engine, defaults.clone());